Move InputLocation/InputLocationTracker/InputSource to velocity templates
This PR moves the InputLocation, InputLocationTracker, and InputSource
classes from maven-api-model to be generated using the existing velocity
templates in src/mdo/java/.

Changes Made:
- Updated Velocity Templates with Maven-specific features controlled by isMavenModel parameter
- Added locationTracking=true and generateLocationClasses=true parameters to maven-api-model configuration
- Removed manually written InputLocation classes from api/maven-api-model/src/main/java/org/apache/maven/api/model/

Benefits:
- Consistency: InputLocation classes are now generated consistently across all Maven modules
- Maintainability: Changes can be made in one place (the templates) rather than maintaining separate implementations
- Feature Parity: All Maven-specific features are preserved through the isMavenModel boolean parameter
- Code Reuse: The same templates can generate both simple and Maven-enhanced versions
